Implements an event-based task-scheduler
It is developed based on the observer model, where the observer is able
to subscribe to a service, such that when the service is updated, the
observer will be informed.

The tasks to perform an update are managed by the TaskScheduler, when it
is time to update, TaskScheduler is triggered and sends the request to
the "smart" thread pool for execution. Once it has been executed, the
observer will be informed.

- The observer inherits from EventObserverInterface
- The subject inherits from SubscriptionServiceInterface

Example implementations to be done.